Rework approach towards builtin resolution
