--- crates/engine/def-map/src/macro_expansion/body.rs
@@ -1,593 +1,0 @@
-//! Body-facing declarative macro expansion.
-//!
-//! Body lowering needs expansion as an input to syntax lowering, but it should not know about the
-//! token-tree and macro-engine crates directly. This facade keeps body-specific frozen def-map
-//! visibility and token conversion next to the def-map query it relies on.
-
-use anyhow::Context as _;
-
-use rg_cfg_eval::CfgEvaluator;
-use rg_ir_model::{
-    BodySource, BuiltinMacroExprKind, DefMapRef, LocalDefRef, ModuleId, ModuleRef, TargetRef,
-};
-use rg_ir_storage::{
-    DefMapQuery, ImportPath, MacroDefinitionData, MacroDefinitionView, PathResolver,
-};
-use rg_macro_runtime::{
-    CfgSelect, ExpansionParseKind, ExpansionSyntax, MacroExpansionRequest, MacroExpansionRuntime,
-    macro_edition,
-};
-use rg_parse::{FileId, Span};
-use rg_std::ExpectedUnique;
-use rg_syntax::{AstNode, Parse, SyntaxNode, ast, utils::normalized_syntax_text};
-use rg_text::Name;
-use rg_tt::TopSubtree;
-use rg_tt::syntax_bridge::{ExpansionSpanMap, SpanFactory, syntax_node_to_token_tree_with_span};
-use rg_workspace::RustEdition;
-
-use crate::DefMapReadTxn;
-
-use super::builtin;
-
-/// Generated body syntax plus the macro-definition origin needed while lowering it.
-pub struct ExpandedBodyMacro<T> {
-    syntax: T,
-    source_map: ExpansionSpanMap,
-    dollar_crate_target: TargetRef,
-}
-
-impl<T> ExpandedBodyMacro<T> {
-    fn new(syntax: T, source_map: ExpansionSpanMap, dollar_crate_target: TargetRef) -> Self {
-        Self {
-            syntax,
-            source_map,
-            dollar_crate_target,
-        }
-    }
-
-    /// Maps generated syntax back to a macro invocation token when that mapping is precise.
-    pub fn source_for_exact_syntax(
-        &self,
-        call_source: BodySource,
-        syntax: &SyntaxNode,
-    ) -> Option<Span> {
-        let range = syntax.text_range();
-
-        // Only accept exact single-token source mappings from the original call. Compound
-        // generated nodes often contain substituted tokens, but the node as a whole was shaped by
-        // the macro transcriber and should stay anchored to the call site.
-        let token = syntax.first_token()?;
-        if token != syntax.last_token()? || token.text_range() != range {
-            return None;
-        }
-
-        let span = self
-            .source_map
-            .span_for_range_in_file(range, call_source.file_id.0)?;
-        let span = Span::from_text_range(span.range);
-        (call_source.span.contains_span(span) && span.len() == u32::from(range.len()))
-            .then_some(span)
-    }
-
-    /// Returns the crate that generated `$crate` paths inside this syntax should resolve to.
-    pub fn dollar_crate_target(&self) -> TargetRef {
-        self.dollar_crate_target
-    }
-}
-
-impl<T: AstNode> ExpandedBodyMacro<T> {
-    /// Splits out typed generated syntax while keeping source context active for descendants.
-    pub fn into_syntax_and_context(self) -> (T, ExpandedBodyMacro<SyntaxNode>) {
-        let Self {
-            syntax,
-            source_map,
-            dollar_crate_target,
-        } = self;
-        let context_syntax = syntax.syntax().clone();
-        (
-            syntax,
-            ExpandedBodyMacro::new(context_syntax, source_map, dollar_crate_target),
-        )
-    }
-}
-
-impl ExpandedBodyMacro<Parse<SyntaxNode>> {
-    fn cast_root_or_child<N: AstNode>(self) -> Option<ExpandedBodyMacro<N>> {
-        let Self {
-            syntax: parse,
-            source_map,
-            dollar_crate_target,
-        } = self;
-        let root = parse.syntax_node();
-        let syntax = N::cast(root.clone()).or_else(|| root.children().find_map(N::cast))?;
-        Some(ExpandedBodyMacro::new(
-            syntax,
-            source_map,
-            dollar_crate_target,
-        ))
-    }
-}
-
-/// Expression-position macro call result after body macro lookup.
-pub enum BodyMacroExprExpansion {
-    /// Declarative macro expansion produced syntax that body lowering should lower recursively.
-    Expanded(ExpandedBodyMacro<ast::Expr>),
-    /// Compiler-provided expression macro that has no declarative transcriber to execute.
-    Builtin(BuiltinMacroExprKind),
-}
-
-/// Tells body macro lookup whether the call came from user-written syntax or generated syntax.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum BodyMacroCallOrigin {
-    /// A macro invocation written in the original source file.
-    Source,
-    /// A macro invocation produced by expanding syntax from a macro definition target.
-    Generated { dollar_crate_target: TargetRef },
-}
-
-impl BodyMacroCallOrigin {
-    fn dollar_crate_target_for_path(self) -> Option<TargetRef> {
-        match self {
-            Self::Source => None,
-            Self::Generated {
-                dollar_crate_target,
-            } => Some(dollar_crate_target),
-        }
-    }
-
-    fn dollar_crate_target_for_expansion(self, caller_target: TargetRef) -> TargetRef {
-        match self {
-            Self::Source => caller_target,
-            Self::Generated {
-                dollar_crate_target,
-            } => dollar_crate_target,
-        }
-    }
-}
-
-/// Expands body macro calls using frozen def-map visibility.
-///
-/// Declarative macros return generated syntax. Expression-position compiler builtins are reported
-/// only after ordinary macro lookup fails so user macros can still shadow builtin names.
-pub struct BodyMacroExpander<'db, 'txn> {
-    def_maps: &'txn DefMapReadTxn<'db>,
-    runtime: MacroExpansionRuntime,
-}
-
-impl<'db, 'txn> BodyMacroExpander<'db, 'txn> {
-    pub fn new(def_maps: &'txn DefMapReadTxn<'db>) -> Self {
-        Self {
-            def_maps,
-            runtime: MacroExpansionRuntime::default(),
-        }
-    }
-
-    /// Resolves one expression-position macro call to generated syntax or a builtin marker.
-    pub fn expand_expr_call(
-        &mut self,
-        target: TargetRef,
-        module: ModuleRef,
-        source: BodySource,
-        origin: BodyMacroCallOrigin,
-        parse_package: &rg_parse::Package,
-        cfg: CfgEvaluator<'_>,
-        call: &ast::MacroCall,
-    ) -> anyhow::Result<Option<BodyMacroExprExpansion>> {
-        let query = DefMapQuery::new(self.def_maps);
-        let Some(resolution) =
-            Self::resolve_call(&query, target, module, source, origin, parse_package, call)?
-        else {
-            return Ok(None);
-        };
-
-        match resolution.resolved {
-            ExpectedUnique::One(resolved) => {
-                let Some(expanded) = self.expand_resolved_invocation(
-                    &resolution.invocation,
-                    resolved,
-                    ExpansionParseKind::Expr,
-                ) else {
-                    return Ok(None);
-                };
-
-                return Ok(expanded
-                    .cast_root_or_child::<ast::Expr>()
-                    .map(BodyMacroExprExpansion::Expanded));
-            }
-            ExpectedUnique::Ambiguous => return Ok(None),
-            ExpectedUnique::Empty => {}
-        }
-
-        if Self::is_cfg_select_builtin(&resolution.path) {
-            return Ok(Self::expand_builtin_cfg_select(
-                &resolution.invocation,
-                cfg,
-                origin.dollar_crate_target_for_expansion(target),
-                ExpansionParseKind::Expr,
-            )
-            .and_then(|expanded| expanded.cast_root_or_child::<ast::Expr>())
-            .map(BodyMacroExprExpansion::Expanded));
-        }
-
-        Ok(
-            builtin::body_expr_kind_from_path(&resolution.path)
-                .map(BodyMacroExprExpansion::Builtin),
-        )
-    }
-
-    /// Expands one statement-position macro call to generated statement-list syntax.
-    pub fn expand_stmt_call(
-        &mut self,
-        target: TargetRef,
-        module: ModuleRef,
-        source: BodySource,
-        origin: BodyMacroCallOrigin,
-        parse_package: &rg_parse::Package,
-        cfg: CfgEvaluator<'_>,
-        call: &ast::MacroCall,
-    ) -> anyhow::Result<Option<ExpandedBodyMacro<ast::MacroStmts>>> {
-        let Some(expanded) = self.expand_call_syntax(
-            target,
-            module,
-            source,
-            origin,
-            parse_package,
-            Some(cfg),
-            call,
-            ExpansionParseKind::Statements,
-        )?
-        else {
-            return Ok(None);
-        };
-
-        Ok(expanded.cast_root_or_child::<ast::MacroStmts>())
-    }
-
-    /// Expands one pattern-position macro call to generated pattern syntax.
-    pub fn expand_pat_call(
-        &mut self,
-        target: TargetRef,
-        module: ModuleRef,
-        source: BodySource,
-        origin: BodyMacroCallOrigin,
-        parse_package: &rg_parse::Package,
-        call: &ast::MacroCall,
-    ) -> anyhow::Result<Option<ExpandedBodyMacro<ast::Pat>>> {
-        let Some(expanded) = self.expand_call_syntax(
-            target,
-            module,
-            source,
-            origin,
-            parse_package,
-            None,
-            call,
-            ExpansionParseKind::Pattern,
-        )?
-        else {
-            return Ok(None);
-        };
-
-        Ok(expanded.cast_root_or_child::<ast::Pat>())
-    }
-
-    /// Expands one type-position macro call to generated type syntax.
-    pub fn expand_type_call(
-        &mut self,
-        target: TargetRef,
-        module: ModuleRef,
-        source: BodySource,
-        origin: BodyMacroCallOrigin,
-        parse_package: &rg_parse::Package,
-        call: &ast::MacroCall,
-    ) -> anyhow::Result<Option<ExpandedBodyMacro<ast::Type>>> {
-        let Some(expanded) = self.expand_call_syntax(
-            target,
-            module,
-            source,
-            origin,
-            parse_package,
-            None,
-            call,
-            ExpansionParseKind::Type,
-        )?
-        else {
-            return Ok(None);
-        };
-
-        Ok(expanded.cast_root_or_child::<ast::Type>())
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    /// Unlike `expand_expr_call`, this does not classify expression builtins. It only expands
-    /// declarative macros plus source-selecting `cfg_select!` statement bodies.
-    fn expand_call_syntax(
-        &mut self,
-        target: TargetRef,
-        module: ModuleRef,
-        source: BodySource,
-        origin: BodyMacroCallOrigin,
-        parse_package: &rg_parse::Package,
-        cfg: Option<CfgEvaluator<'_>>,
-        call: &ast::MacroCall,
-        parse_kind: ExpansionParseKind,
-    ) -> anyhow::Result<Option<ExpandedBodyMacro<Parse<SyntaxNode>>>> {
-        let query = DefMapQuery::new(self.def_maps);
-        let Some(resolution) =
-            Self::resolve_call(&query, target, module, source, origin, parse_package, call)?
-        else {
-            return Ok(None);
-        };
-        let resolved = match resolution.resolved {
-            ExpectedUnique::One(resolved) => resolved,
-            ExpectedUnique::Ambiguous => return Ok(None),
-            ExpectedUnique::Empty => {
-                if parse_kind == ExpansionParseKind::Statements
-                    && Self::is_cfg_select_builtin(&resolution.path)
-                {
-                    let Some(cfg) = cfg else {
-                        return Ok(None);
-                    };
-                    return Ok(Self::expand_builtin_cfg_select(
-                        &resolution.invocation,
-                        cfg,
-                        origin.dollar_crate_target_for_expansion(target),
-                        parse_kind,
-                    ));
-                }
-                return Ok(None);
-            }
-        };
-
-        Ok(self.expand_resolved_invocation(&resolution.invocation, resolved, parse_kind))
-    }
-
-    fn expand_builtin_cfg_select(
-        invocation: &BodyMacroInvocation,
-        cfg: CfgEvaluator<'_>,
-        dollar_crate_target: TargetRef,
-        parse_kind: ExpansionParseKind,
-    ) -> Option<ExpandedBodyMacro<Parse<SyntaxNode>>> {
-        let cfg_select = CfgSelect::parse(&invocation.args)?;
-        let arm = cfg_select
-            .arms()
-            .iter()
-            .find(|arm| cfg.is_predicate_enabled(&arm.predicate))?;
-        let ExpansionSyntax { parse, span_map } =
-            ExpansionSyntax::from_token_tree(arm.payload.clone(), parse_kind);
-
-        Some(ExpandedBodyMacro::new(parse, span_map, dollar_crate_target))
-    }
-
-    fn is_cfg_select_builtin(path: &ImportPath) -> bool {
-        matches!(
-            builtin::BuiltinMacroDisposition::from_path(path),
-            Some(builtin::BuiltinMacroDisposition::CfgSelect)
-        )
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    fn resolve_call<'a>(
-        query: &'a DefMapQuery<&DefMapReadTxn<'_>>,
-        target: TargetRef,
-        module: ModuleRef,
-        source: BodySource,
-        origin: BodyMacroCallOrigin,
-        parse_package: &rg_parse::Package,
-        call: &ast::MacroCall,
-    ) -> anyhow::Result<Option<BodyMacroCallResolution<'a>>> {
-        let invocation = match BodyMacroInvocation::from_ast(
-            source.file_id,
-            source.span,
-            parse_package.edition(),
-            call,
-        ) {
-            Some(invocation) => invocation,
-            None => return Ok(None),
-        };
-        // `$crate` in a generated macro call belongs to the macro definition that produced this
-        // syntax. Source calls do not have such a definition context, so `$crate` remains invalid
-        // for them here.
-        // TODO: soft hack, we are not inside of macro resolution context here, so we use this
-        // for the lack of better method; probably we should get rid of `ImportPath` whatsoever
-        // (it exists for historical reasons mostly, and it's equivalent to `Path`) and introduce
-        // appropriate constructors.
-        let Some(path) = ImportPath::from_macro_path_text(
-            invocation.path_text(),
-            origin.dollar_crate_target_for_path(),
-        ) else {
-            return Ok(None);
-        };
-        let resolved = Self::resolve_macro_definition(query, target, module, &path)
-            .context("while attempting to resolve body macro call")?;
-
-        Ok(Some(BodyMacroCallResolution {
-            invocation,
-            path,
-            resolved,
-        }))
-    }
-
-    fn expand_resolved_invocation(
-        &mut self,
-        invocation: &BodyMacroInvocation,
-        resolved: MacroDefinitionView<'_>,
-        parse_kind: ExpansionParseKind,
-    ) -> Option<ExpandedBodyMacro<Parse<SyntaxNode>>> {
-        let request = invocation.expansion_request(resolved.def_ref, resolved.data, parse_kind);
-        let Some(ExpansionSyntax { parse, span_map }) = self.runtime.expand_now(request) else {
-            return None;
-        };
-
-        Some(ExpandedBodyMacro::new(
-            parse,
-            span_map,
-            resolved.data.dollar_crate_target,
-        ))
-    }
-
-    fn resolve_macro_definition<'a>(
-        query: &'a DefMapQuery<&DefMapReadTxn<'_>>,
-        target: TargetRef,
-        module: ModuleRef,
-        path: &ImportPath,
-    ) -> anyhow::Result<ExpectedUnique<MacroDefinitionView<'a>>> {
-        // Body expansion is target-local. Synthetic body modules resolve through their semantic
-        // fallback before reaching this facade.
-        let Some(module_target) = module.origin.as_target_ref() else {
-            return Ok(ExpectedUnique::Empty);
-        };
-        if module_target != target {
-            return Ok(ExpectedUnique::Empty);
-        }
-
-        if let Some(name) = path.relative_single_name() {
-            return Self::resolve_single_name_macro(query, target, module.module, name);
-        }
-
-        let bindings = PathResolver::new(query)
-            .macro_bindings(target, module.module, path)
-            .context("while attempting to resolve qualified body macro path")?;
-        let mut macros = ExpectedUnique::new();
-        for binding in bindings {
-            // Path resolution may return duplicate bindings to the same macro definition; collapse
-            // those while still treating different definitions as ambiguous.
-            if let Some(macro_) = query
-                .macro_definition_view(binding.def)
-                .context("while attempting to fetch body macro definition")?
-            {
-                macros.push(macro_);
-            }
-        }
-
-        Ok(macros)
-    }
-
-    fn resolve_single_name_macro<'a>(
-        query: &'a DefMapQuery<&DefMapReadTxn<'_>>,
-        target: TargetRef,
-        module: ModuleId,
-        name: &Name,
-    ) -> anyhow::Result<ExpectedUnique<MacroDefinitionView<'a>>> {
-        let importing_module = ModuleRef {
-            origin: DefMapRef::Target(target),
-            module,
-        };
-        let mut module_scope_modules = Vec::new();
-        let mut current = Some(module);
-
-        // Note: Body macro expansion intentionally uses the frozen module graph as an approximation
-        // of item-position macro visibility. We do not model body-specific textual ordering here:
-        // doing so would require cross-body statement order and nested body macro scope tracking,
-        // while real projects overwhelmingly use macros that are already module-visible.
-        while let Some(module) = current {
-            let module_ref = ModuleRef {
-                origin: DefMapRef::Target(target),
-                module,
-            };
-            module_scope_modules.push(module_ref);
-
-            current = query
-                .module_data(module_ref)
-                .context("while attempting to fetch parent module for body macro lookup")?
-                .and_then(|module| module.parent);
-        }
-
-        let bindings = PathResolver::new(query)
-            .visible_unqualified_macro_bindings(importing_module, module_scope_modules, name)
-            .context("while attempting to resolve unqualified body macro")?;
-
-        let module_scope = Self::resolve_macro_bindings(query, bindings.module_scope)?;
-        if !module_scope.is_empty() {
-            return Ok(module_scope);
-        }
-
-        Self::resolve_macro_bindings(query, bindings.standard_prelude)
-    }
-
-    fn resolve_macro_bindings<'a>(
-        query: &'a DefMapQuery<&DefMapReadTxn<'_>>,
-        bindings: Vec<rg_ir_storage::ScopeBinding>,
-    ) -> anyhow::Result<ExpectedUnique<MacroDefinitionView<'a>>> {
-        let mut resolved = ExpectedUnique::new();
-
-        for binding in bindings {
-            if let Some(macro_) = query
-                .macro_definition_view(binding.def)
-                .context("while attempting to fetch body macro definition")?
-            {
-                resolved.push(macro_);
-            }
-        }
-
-        Ok(resolved)
-    }
-}
-
-/// Body macro call after parsing and macro lookup.
-struct BodyMacroCallResolution<'a> {
-    invocation: BodyMacroInvocation,
-    path: ImportPath,
-    resolved: ExpectedUnique<MacroDefinitionView<'a>>,
-}
-
-/// Body-specific adapter from parsed macro-call syntax to runtime expansion input.
-///
-/// Item-position calls are already lowered by item-tree before def-map expansion sees them. Bodies
-/// arrive here as `ast::MacroCall`, so this private adapter keeps the AST and token-tree conversion
-/// next to the body visibility policy instead of making `rg_macro_runtime` depend on parsed AST.
-struct BodyMacroInvocation {
-    path_text: String,
-    args: TopSubtree,
-    call_file_id: FileId,
-    call_span: Span,
-    call_edition: RustEdition,
-}
-
-impl BodyMacroInvocation {
-    fn from_ast(
-        file_id: FileId,
-        span: Span,
-        edition: RustEdition,
-        call: &ast::MacroCall,
-    ) -> Option<Self> {
-        let path_text = call.path().map(|path| normalized_syntax_text(&path))?;
-        let args = call.token_tree()?;
-
-        let span_factory = SpanFactory::new(
-            u32::try_from(file_id.0).expect("file id should fit macro span storage"),
-            macro_edition(edition),
-        );
-        let args =
-            syntax_node_to_token_tree_with_span(&args, &mut |range| span_factory.span_for(range));
-
-        Some(Self {
-            path_text,
-            args,
-            call_file_id: file_id,
-            call_span: span,
-            call_edition: edition,
-        })
-    }
-
-    fn path_text(&self) -> &str {
-        &self.path_text
-    }
-
-    fn expansion_request<'a>(
-        &'a self,
-        def_ref: LocalDefRef,
-        definition: &'a MacroDefinitionData,
-        parse_kind: ExpansionParseKind,
-    ) -> MacroExpansionRequest<'a> {
-        MacroExpansionRequest {
-            def_ref,
-            definition,
-            path_text: &self.path_text,
-            args: &self.args,
-            call_file_id: self.call_file_id,
-            call_span: self.call_span,
-            call_edition: self.call_edition,
-            parse_kind,
-        }
-    }
-}
