+++ crates/engine/def-map/src/macro_expansion/builtin.rs
@@ -1,58 +1,21 @@
-//! Shared classification for compiler-provided macro names.
+//! Shared fallback classification for compiler-provided macro names.
 //!
-//! Macro lookup should always prefer user definitions first. This classifier is only consulted
-//! after normal macro resolution fails, so local `macro_rules! format_args` can still shadow the
-//! compiler builtin.
+//! Macro lookup should always prefer user definitions first. Resolved definitions carry builtin
+//! identity directly when item-tree saw `#[rustc_builtin_macro]`; the path classifier below remains
+//! a fallback for names that do not resolve through the sysroot yet.
 
-use rg_ir_model::{BuiltinMacroExprKind, PathSegment};
+use rg_ir_model::{PathSegment, items::BuiltinMacroKind};
 use rg_ir_storage::ImportPath;
 use rg_text::Name;
 
-/// Known builtin macro call that should not be resolved as a user macro.
-pub(crate) enum BuiltinMacroDisposition {
-    /// The builtin cannot add module-scope definitions, so def-map can safely ignore it.
-    IgnoredByDefMap,
-    /// The builtin selects one item stream from cfg predicates.
-    CfgSelect,
-    /// The builtin splices another source file into the caller's item stream.
-    Include,
-    /// The builtin can affect item collection or requires dedicated compiler-like handling.
-    Unsupported,
-}
-
-impl BuiltinMacroDisposition {
-    /// Classifies builtin macros that are known even when no user macro binding resolves.
-    ///
-    /// We intentionally take a small shortcut here. Unqualified builtins and `std`/`core`-qualified
-    /// builtin-shaped paths cover the realistic call sites, while a fully resolution-aware builtin
-    /// prefix model would add a lot of complexity for rare local `std`/`core` shadowing cases.
-    pub(crate) fn from_path(path: &ImportPath) -> Option<Self> {
-        let name = macro_name_from_path(path)?;
-
-        match name.as_str() {
-            // Expression, diagnostic, or assembly builtins do not contribute named items to def-map.
-            // Body lowering can synthesize values/types for the expression-like subset.
-            "asm" | "cfg" | "column" | "compile_error" | "concat" | "env" | "file"
-            | "format_args" | "format_args_nl" | "global_asm" | "include_bytes" | "include_str"
-            | "line" | "llvm_asm" | "module_path" | "option_env" | "stringify" => {
-                Some(Self::IgnoredByDefMap)
-            }
-
-            "cfg_select" => Some(Self::CfgSelect),
-            "include" => Some(Self::Include),
-
-            // `concat_idents!` has token-shaping behavior that is better handled by a dedicated
-            // builtin implementation.
-            "concat_idents" => Some(Self::Unsupported),
-            _ => None,
-        }
-    }
-}
-
-/// Returns the Body IR expression shape for builtin macros that are useful inside bodies.
-pub(crate) fn body_expr_kind_from_path(path: &ImportPath) -> Option<BuiltinMacroExprKind> {
+/// Classifies builtin macros that are known even when no user macro binding resolves.
+///
+/// We intentionally take a small shortcut here. Unqualified builtins and `std`/`core`-qualified
+/// builtin-shaped paths cover the realistic call sites, while a fully resolution-aware builtin
+/// prefix model would add a lot of complexity for rare local `std`/`core` shadowing cases.
+pub(crate) fn kind_from_path(path: &ImportPath) -> Option<BuiltinMacroKind> {
     let name = macro_name_from_path(path)?;
-    BuiltinMacroExprKind::from_macro_name(name.as_str())
+    BuiltinMacroKind::from_known_name(name.as_str())
 }
 
 // TODO: Shortcut/heuristic macro resolution for builtins: either direct path,
