+++ crates/engine/ir-model/src/items/macro_item/builtin.rs
@@ -9,7 +9,48 @@ use rg_parse::FileId;
 use rg_std::{MemorySize, Shrink};
 use wincode::{SchemaRead, SchemaWrite};
 
+use crate::BuiltinMacroExprKind;
+
 use super::super::ItemTreeId;
+
+/// Compiler-provided macro definition selected through normal macro resolution.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, SchemaRead, SchemaWrite, MemorySize, Shrink)]
+#[memsize(leaf)]
+#[shrink(leaf)]
+pub enum BuiltinMacroKind {
+    /// Builtin expression macro that Body IR can represent without declarative expansion.
+    Expr(BuiltinMacroExprKind),
+    /// Builtin that selects one source stream from cfg predicates.
+    CfgSelect,
+    /// Builtin that splices another source file into the caller.
+    Include,
+    /// Builtin that does not contribute item definitions to def-map.
+    IgnoredByDefMap,
+    /// Builtin that needs dedicated support before it can be expanded.
+    Unsupported,
+}
+
+impl BuiltinMacroKind {
+    /// Classify a known compiler builtin name.
+    pub fn from_known_name(name: &str) -> Option<Self> {
+        if let Some(kind) = BuiltinMacroExprKind::from_macro_name(name) {
+            return Some(Self::Expr(kind));
+        }
+
+        match name {
+            "asm" | "compile_error" | "global_asm" | "llvm_asm" => Some(Self::IgnoredByDefMap),
+            "cfg_select" => Some(Self::CfgSelect),
+            "include" => Some(Self::Include),
+            "concat_idents" => Some(Self::Unsupported),
+            _ => None,
+        }
+    }
+
+    /// Classify a definition explicitly marked by rustc as compiler-provided.
+    pub fn from_rustc_builtin_macro_name(name: &str) -> Self {
+        Self::from_known_name(name).unwrap_or(Self::Unsupported)
+    }
+}
 
 /// Source-like builtin payload discovered during item-tree lowering.
 #[derive(Debug, Clone, PartialEq, Eq, SchemaRead, SchemaWrite, MemorySize, Shrink)]
